// src/app/model/pago.ts
export class Pago{
    id_pago: number = 0;
    nombreapellido_pago: string = "";
    correo_pago: string = "";
    direccion_pago: string = "";
    distrito_pago: string = "";
    telefono_pago: number = 0;
    raza_pago: string = "";
    tarjeta_pago: string = "";
    propietario_pago: string = "";
    numTarjeta_pago: number = 0;
    mesexpiracion_pago: string = "";
    añoexpiracion_pago: number = 0;
    cvv_pago: number = 0;
}